change RQ to alphanumeric field (e.g. "low")
After a long back-and-forth exchange with Chrome & Opera folks, where we
tried to converge on some "common values" that we can agree on and
recommend to developers, we realized that we may be overloading a single
hint: we want to account for user preference, bandwidth, content type,
and so on. This makes it very hard to interpret the hint and will lead
to unnecessary fragmentation.

With that in mind, the intent here is to simplify the hint and
comunicate a single bit of information: user preference for "lower
resource quality", which itself may have multiple keyword values.

For example, in Chrome, when the user enables "data savings" the browser
may advertise "low" as RQ value to indicate that the user wants to save
data. Alternatively, in Opera, the user can enable a similar option but
may also be able to set one of multiple compression levels (e.g. low,
medium, high).