first draft for Alternate-Content-Length
A mechanism to help the client estimate byte savings when advertised
hints are used by the server to select and deliver an optimized
response.

E.g. an optimizing proxy needs a way to convey the content length of the
alternate (unmodified) response body to allow the user agent to track
and report compression savings. Today, Chrome uses
X-Original-Content-Length for this, and Opera has a similar hader.